Moved general-purpose post-processors to post_process module.
